requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
cryptography>=41.0.0
python-dotenv>=1.0.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0
datasets>=2.14.0
evaluate>=0.4.0
spotipy
spacy
